Login und chatfenster in einen einzelnen Controller verpackt. Scenes müssen noch nach login geswitched werden. Dynamische Nutzerliste fehlt noch
