--- serverclient/ClientControllerTest.java
@@ -1,41 +1,0 @@
-package serverclient;
-
-import java.io.FileInputStream;
-import java.io.IOException;
-
-import javafx.application.Application;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Scene;
-import javafx.scene.layout.AnchorPane;
-import javafx.stage.Stage;
-
-public class ClientControllerTest extends Application
-{
-    public static void main(String[] args)
-    {
-        Application.launch(args);
-    }
-
-    @Override
-    public void start(Stage stage) throws IOException
-    {
-        // Create the FXMLLoader 
-        FXMLLoader loader = new FXMLLoader();
-        // Path to the FXML File
-        String fxmlDocPath = "/home/sophie/Documents/Programmierpraktikum/serverclient/ClientLogin.fxml";
-        FileInputStream fxmlStream = new FileInputStream(fxmlDocPath);
-
-        // Create the Pane and all Details
-        AnchorPane root = (AnchorPane) loader.load(fxmlStream);
-
-        // Create the Scene
-        Scene scene = new Scene(root);
-        // Set the Scene to the Stage
-        stage.setScene(scene);
-        // Set the Title to the Stage
-        stage.setTitle("First Controller Babysteps");
-        // Display the Stage
-        stage.show();
-    }
-
-}
